requirements.txt
Django==1.8.2
Jinja2==2.7.3
Markdown==2.5.2
MarkupSafe==0.23
PyYAML==3.11
backports.ssl-match-hostname==3.4.0.2
certifi==2015.04.28
click==4.0
django-filter==0.9.2
django-guardian==1.2.5
djangorestframework==3.1.3
flake8==2.4.0
ghp-import==0.4.1
livereload==2.4.0
mccabe==0.3
mkdocs==0.13.2
pep8==1.5.7
pkginfo==1.2.1
py==1.4.28
pyflakes==0.8.1
pytest==2.6.4
pytest-django==2.8.0
requests==2.7.0
six==1.9.0
tornado==4.2
transifex-client==0.10
twine==1.4.0
wheel==0.24.0
wsgiref==0.1.2
